Add dagger tests
